Separation of classes into separate files
- Files directory creation
- Images directory creation
- Builder directory creation
- Separate classes into several files
- Correct buttons
- Change colors
- Correction of lists usable only with Bladebit
